Git malina

+++ src/main/java/razdwatrzy/zzpj/crypto/ICipher.java
@@ -4,6 +4,6 @@ import org.springframework.stereotype.Service;
 
 
 public interface ICipher {
-    byte[] encode(int i) throws Exception;
-    int decode(byte[] s) throws Exception;
+    byte[] encode(long i) throws Exception;
+    long decode(byte[] s) throws Exception;
 }
